+++ deliverables/sync_filled_from_docx.py
@@ -52,7 +52,7 @@ FIGS = [
     "model_comparison_components_vv_L24_2150.png",
     "model_comparison_components_vv_L24_2300.png",
     "future_components_vv_L24_joint.png",
-    "vv_gsic_ladrillo_2300.png",       # Ladrillo-only since 9/11 comment [11]; was vv_gsic_wr_vs_ladrillo_2300.png
+    "vv_gsic_ladrillo_L24_2300.png",   # Ladrillo-only since 9/11 comment [11]; tag added to the name 09-20
     "vv_responsiveness_L24.png",       # FIG 7, 09-12
     "model_comparison_components_L24_2100.png",
     "model_comparison_components_L24_2300.png",